requirements.txt
numpy
opencv-python
scikit-image
scikit-learn
matplotlib
tensorflow-hub
seaborn
pandas
Pillow
